\name{adminGUI}
\alias{adminGUI}
\title{adminGUI}
\description{Graphical admin interface for interactive and probably statistical courses}
\usage{adminGUI(pwd="Methodik2009", evaluierung = FALSE, enableAll=TRUE) }
\arguments{
	\item{pwd}{admin-GUI may be started with password only}
  	\item{evaluierung}{should GUI be started in evaluation mode, default=FALSE}
  	\item{enableAll}{should GUI be started with all available options, default=FALSE}
}
\details{To be written}
\value{Graphical Admin Interface}
\author{Matthias Templ, Gerlinde Dinges, Alexander Kowarik, Bernhard Meindl}
\examples{
	\dontrun{
		adminGUI()
	}	
}
\keyword{methods}