\name{ExitButton}
\alias{ExitButton}
\title{ExitButton}
\description{returns an Exit-Button in a frame called \emph{name} within frame \emph{frame} in tk-window \emph{window}.}
\usage{ExitButton(window=tt, frame=MainFrame, name, text="Close")}
\arguments{
	\item{window}{the tk-toplevel() in which the ExitButton should be drawn.}
	\item{frame}{the frame in which the ExitButton should be drawn.}
	\item{name}{the name of the frame that is created and which includes the ExitButton}
	\item{text}{the text under the ExitButton}
}
\details{internal function only}
\author{Alexander Kowarik, Bernhard Meindl}
\keyword{methods}