\name{runGUI}
\alias{runGUI}
\title{runGUI}
\description{Graphical user interface for interactive and probably statistical courses}
\usage{runGUI()}
\arguments{
}
\details{To be written}
\value{Graphical User Interface}
\author{Matthias Templ, Gerlinde Dinges, Alexander Kowarik, Bernhard Meindl}
\examples{
	\dontrun{
		runGUI()
	}	
}
\keyword{methods}