\name{setFont}
\alias{setFont}
\title{setFont}
\description{sets tcltk font to be used in widgets.}
\usage{setFont(family="tahoma", size, italic=FALSE, bold=FALSE)}
\arguments{
	\item{family}{the font family that should be used, default is tahoma.}
	\item{size}{the font size. Specifying a numeric value is possible just as specifying either small, normal, large, extralarge or huge.}
	\item{italic}{should the font be italic, default=FALSE}
	\item{bold}{should the font be bold, default=FALSE}
}
\details{internal function only}
\author{Alexander Kowarik, Bernhard Meindl}
\keyword{methods}