\name{showPDF}
\alias{showPDF}
\title{showPDF}
\description{this function opens a given pdf relative to the doc-path of package-installation in the standard pdf-viewer}
\usage{showPDF(filename) }
\arguments{
	\item{filename}{the filename of the pdf to be opened.}
}
\details{internal function only}
\author{Alexander Kowarik, Bernhard Meindl}
\keyword{methods}