\name{aTassign}
\alias{aTassign}
\alias{aTget}
\alias{aTexists}
\alias{aTrm}
\title{aTassign}
\description{assign/delete/check/retrieve an object to/in/from the namespace adminTGUIenv.}
\usage{aTassign(x,value)}
\usage{aTget(x)}
\usage{aTrm(x)}
\usage{aTexists(x)}
\arguments{
	\item{x}{the name of the object to be created/loaded/checked/removed in/from namespace adminTGUIenv.}
	\item{value}{the value/object to be assigned.}
}
\details{internal function only}
\author{Alexander Kowarik, Bernhard Meindl}
\keyword{methods}