\name{runGUI}
\alias{runGUI}
\title{runGUI}
\description{Graphical user interface for interactive and probably statistical courses}
\usage{runGUI(evaluierung = FALSE)}
\arguments{
  \item{evaluierung}{should GUI be started in evaluation mode, default=FALSE}
}
\details{To be written}
\value{Graphical User Interface}
\author{Matthias Templ, Gerlinde Dinges, Alexander Kowarik, Bernhard Meindl}
\examples{
	\dontrun{
		runGUI()
	}	
}
\keyword{methods}