\name{Tassign}
\alias{Tassign}
\alias{Tget}
\alias{Texists}
\alias{Trm}
\alias{Tclear}
\title{Tassign}
\description{assign/delete/check/retrieve an object to/in/from the namespace TGUIenv.}
\usage{Tassign(x,value)
Tget(x)
Trm(x)
Texists(x)
Tclear()}

\arguments{
	\item{x}{the name of the object to be created/loaded/checked/removed in/from namespace TGUIenv.}
	\item{value}{the value/object to be assigned.}
}
\details{internal function only}
\author{Alexander Kowarik, Bernhard Meindl}
\keyword{methods}